WIP Browser build
* use rollup
* ES6 cause, it is lovely
