+++ .babelrc
@@ -1,0 +1,4 @@
+// .babelrc
+{
+  "presets": [ "es2015-rollup" ]
+}
