+++ index.config.js
@@ -1,0 +1,20 @@
+import nodeResolve from 'rollup-plugin-node-resolve';
+import babel from 'rollup-plugin-babel';
+import commonjs from 'rollup-plugin-commonjs';
+
+export default {
+  entry: 'index.js',
+  moduleName: 'heimdall',
+  plugins: [
+    babel({
+      exclude: 'node_modules/**'
+    }),
+    nodeResolve({ jsnext: true, main: true }),
+    commonjs({ include: 'node_modules/**' })
+  ],
+  targets: [
+    { dest: 'dist/bundle.cjs.js', format: 'cjs' },
+    { dest: 'dist/bundle.umd.js', format: 'umd' },
+    { dest: 'dist/bundle.es.js', format: 'es' },
+  ]
+}
