+++ index.html
@@ -1,0 +1,1 @@
+<script src="dist/bundle.umd.js"></script>
