+++ package.json
@@ -4,7 +4,8 @@
   "description": "Structured instrumentation library",
   "main": "index.js",
   "scripts": {
-    "test": "mocha tests/",
+    "build": "rollup --no-strict -c test.config.js && rollup -c index.config.js",
+    "test": "npm run build && mocha dist/tests/bundle.cjs",
     "test:debug": "mocha --no-timeouts debug tests/"
   },
   "repository": {
@@ -20,13 +21,23 @@
   "keywords": [
     "javascript"
   ],
+  "files": [
+    "src",
+    "dist"
+  ],
   "devDependencies": {
+    "babel-preset-es2015-rollup": "^1.1.1",
     "broccoli": "^0.16.9",
     "chai": "^3.2.0",
     "chai-as-promised": "^5.1.0",
     "chai-files": "^1.2.0",
     "mocha": "^2.2.5",
-    "mocha-jshint": "~2.2.3"
+    "mocha-jshint": "~2.2.3",
+    "rollup": "^0.34.1",
+    "rollup-plugin-babel": "^2.6.1",
+    "rollup-plugin-buble": "^0.12.1",
+    "rollup-plugin-commonjs": "^3.3.1",
+    "rollup-plugin-node-resolve": "^1.7.1"
   },
   "dependencies": {
     "rsvp": "^3.2.1"
