+++ src/cookie.js
@@ -1,37 +1,36 @@
-module.exports = Cookie;
-function Cookie(node, heimdall) {
-  this._node = node;
-  this._restoreNode = node.parent;
-  this._heimdall = heimdall;
-  this._stopped = false;
-}
+export default class Cookie {
+  constructor(node, heimdall) {
+    this._node = node;
+    this._restoreNode = node.parent;
+    this._heimdall = heimdall;
+    this._stopped = false;
+  }
 
-Object.defineProperty(Cookie.prototype, 'stats', {
-  get: function() {
+  get stats() {
     return this._node.stats.own;
   }
-});
 
-Cookie.prototype.stop = function() {
-  var monitor;
+  stop() {
+    let monitor;
 
-  if (this._heimdall.current !== this._node) {
-    throw new TypeError('cannot stop: not the current node');
-  } else if (this.stopped === true) {
-    throw new TypeError('cannot stop: already stopped');
+    if (this._heimdall.current !== this._node) {
+      throw new TypeError('cannot stop: not the current node');
+    } else if (this.stopped === true) {
+      throw new TypeError('cannot stop: already stopped');
+    }
+
+    this._stopped = true;
+    this._heimdall._recordTime();
+    this._heimdall._session.current = this._restoreNode;
   }
 
-  this._stopped = true;
-  this._heimdall._recordTime();
-  this._heimdall._session.current = this._restoreNode;
-};
+  resume() {
+    if (this._stopped === false) {
+      throw new TypeError('cannot resume: not stopped');
+    }
 
-Cookie.prototype.resume = function() {
-  if (this._stopped === false) {
-    throw new TypeError('cannot resume: not stopped');
+    this._stopped = false;
+    this._restoreNode = this._heimdall.current;
+    this._heimdall._session.current = this._node;
   }
-
-  this._stopped = false;
-  this._restoreNode = this._heimdall.current;
-  this._heimdall._session.current = this._node;
-};
+}
